feat(workspace): implement equal stacks algorithm
This commit implements the equal stacks algorithm.  The
algorithm takes three stacks of integers as input and
returns the maximum possible height of the stacks such
that all of the stacks have the same height.  The height
of a stack is the sum of the integers in the stack. The
algorithm works by repeatedly removing the top element
from the tallest stack until all of the stacks have the
same height.